commit v10.0.0

+++ cupcake2/tofu2/create_seed.py
@@ -15,7 +15,7 @@ def create_seed_n_batch_files(input='isoseq_flnc.fasta', fasta_d=None, seed_file
     n = len(lens)
 
     # start at 1% of the data
-    starting_seed_index = n * 1 / 100
+    starting_seed_index = int(n * 1 / 100)
     good = [x[0] for x in lens[starting_seed_index:starting_seed_index+num_seqs_per_batch]]
     write_seqids_to_fasta(good, seed_filename, fasta_d)
